refactor(data-structures): move transaction_tag from validations to chain.rs
